Implement Task 8: Modular Architecture
Restructured the codebase from a monolithic 943-line object_detection.py
into a clean, modular architecture with clear separation of concerns.

## New Structure:

models/
├── __init__.py
├── yolov3.py          # YOLOv3 architecture (224 lines)
└── weight_loader.py   # Darknet weight reader (118 lines)

detection/
├── __init__.py
├── detector.py        # Lazy-loading detector (180 lines)
├── preprocessing.py   # Preprocessing & decoding (213 lines)
└── violation_checker.py  # Violation detection (229 lines)

gui/
└── __init__.py        # GUI package (future refactoring)

## Key Changes:

1. **Separated Model Architecture**: YOLOv3 model and weight loading
   isolated in models/ package

2. **Detection Pipeline**: Split preprocessing, detection, and violation
   checking into focused modules

3. **Backward Compatibility**: object_detection.py maintained as
   deprecation wrapper - existing code still works

4. **Updated Imports**: main.py now uses modular imports from
   detection/ and models/ packages

5. **Documentation**: Added comprehensive ARCHITECTURE.md explaining
   the new structure, migration guide, and benefits

## Benefits:

- ✅ Separation of concerns (each module has single responsibility)
- ✅ Improved testability (modules can be tested independently)
- ✅ Better organization (related functions grouped together)
- ✅ Reusability (modules can be imported independently)
- ✅ Maintainability (changes isolated to specific modules)
- ✅ Scalability (easy to add new features)

## Testing:

- ✅ All imports working correctly
- ✅ Backward compatibility verified
- ✅ No breaking changes to existing code
