+++ gui/__init__.py
@@ -1,0 +1,10 @@
+"""
+GUI Interface for Traffic Violation Detection System
+
+This package contains the graphical user interface components.
+"""
+
+# GUI module will be refactored in a future update
+# Currently using Project-GUI.py
+
+__all__ = []
