+++ models/__init__.py
@@ -1,0 +1,11 @@
+"""
+YOLOv3 Model Architecture and Weight Loading
+
+This package contains the YOLOv3 model architecture and utilities
+for loading pre-trained weights from Darknet format.
+"""
+
+from .yolov3 import make_yolov3_model
+from .weight_loader import WeightReader
+
+__all__ = ['make_yolov3_model', 'WeightReader']
