+++ object_detection.py
@@ -1,943 +1,57 @@
-import numpy as np
-from tensorflow.keras.layers import Conv2D, Input, BatchNormalization, LeakyReLU, ZeroPadding2D, UpSampling2D, Add, Concatenate
-from tensorflow.keras.models import Model
-import struct
-import cv2
-import time
-import logging
-from utils.geometry import bounding_box_intersects_line
-import config
-
-# ============================================================================
-# GPU Configuration (must be done before model loading)
-# ============================================================================
-
-# Configure GPU before any TensorFlow operations
-_gpus = config.configure_gpu()
-if _gpus:
-    logging.info(f"YOLOv3 will use GPU acceleration: {_gpus}")
-else:
-    logging.info("YOLOv3 will run on CPU")
-
-class WeightReader:
-    """
-    Reader for loading YOLOv3 pre-trained weights from Darknet format.
-
-    This class reads binary weight files in the Darknet format (.weights)
-    and loads them into a Keras model. Handles version compatibility and
-    proper weight ordering for convolutional and batch normalization layers.
-
-    Attributes:
-        offset (int): Current position in the weight array
-        all_weights (np.ndarray): All weights loaded from file as float32
-
-    Example:
-        >>> model = make_yolov3_model()
-        >>> weight_reader = WeightReader('yolov3.weights')
-        >>> weight_reader.load_weights(model)
-        >>> # Model is now ready for inference
-    """
-    def __init__(self, weight_file):
-        """
-        Initialize weight reader and load weights from file.
-
-        Args:
-            weight_file (str): Path to Darknet weights file (.weights)
-
-        Raises:
-            FileNotFoundError: If weight file doesn't exist
-            struct.error: If file format is invalid
-        """
-        with open(weight_file, 'rb') as w_f:
-            major,    = struct.unpack('i', w_f.read(4))
-            minor,    = struct.unpack('i', w_f.read(4))
-            revision, = struct.unpack('i', w_f.read(4))
-
-            if (major*10 + minor) >= 2 and major < 1000 and minor < 1000:
-                w_f.read(8)
-            else:
-                w_f.read(4)
-
-            transpose = (major > 1000) or (minor > 1000)
-
-            binary = w_f.read()
-
-        self.offset = 0
-        self.all_weights = np.frombuffer(binary, dtype='float32')
-
-    def read_bytes(self, size):
-        """
-        Read specified number of weights from current position.
-
-        Args:
-            size (int): Number of weight values to read
-
-        Returns:
-            np.ndarray: Array of weight values
-        """
-        self.offset = self.offset + size
-        return self.all_weights[self.offset-size:self.offset]
-
-    def load_weights(self, model):
-        """
-        Load pre-trained weights into YOLOv3 Keras model.
-
-        Iterates through all 106 convolutional layers in YOLOv3 and loads
-        weights from the Darknet format, handling both batch normalization
-        and bias parameters correctly.
-
-        Args:
-            model (keras.Model): YOLOv3 model created by make_yolov3_model()
-
-        Note:
-            Layers 81, 93, and 105 are detection layers without batch norm.
-            Weights are transposed from Darknet [out, in, h, w] to
-            Keras [h, w, in, out] format.
-
-        Example:
-            >>> model = make_yolov3_model()
-            >>> reader = WeightReader('yolov3.weights')
-            >>> reader.load_weights(model)
-        """
-        for i in range(106):
-            try:
-                conv_layer = model.get_layer('conv_' + str(i))
-                logging.debug(f"Loading weights of convolution #{i}")
-
-                if i not in [81, 93, 105]:
-                    norm_layer = model.get_layer('bnorm_' + str(i))
-
-                    size = np.prod(norm_layer.get_weights()[0].shape)
-
-                    beta  = self.read_bytes(size) # bias
-                    gamma = self.read_bytes(size) # scale
-                    mean  = self.read_bytes(size) # mean
-                    var   = self.read_bytes(size) # variance
-
-                    weights = norm_layer.set_weights([gamma, beta, mean, var])
-
-                if len(conv_layer.get_weights()) > 1:
-                    bias   = self.read_bytes(np.prod(conv_layer.get_weights()[1].shape))
-                    kernel = self.read_bytes(np.prod(conv_layer.get_weights()[0].shape))
-
-                    kernel = kernel.reshape(list(reversed(conv_layer.get_weights()[0].shape)))
-                    kernel = kernel.transpose([2,3,1,0])
-                    conv_layer.set_weights([kernel, bias])
-                else:
-                    kernel = self.read_bytes(np.prod(conv_layer.get_weights()[0].shape))
-                    kernel = kernel.reshape(list(reversed(conv_layer.get_weights()[0].shape)))
-                    kernel = kernel.transpose([2,3,1,0])
-                    conv_layer.set_weights([kernel])
-            except ValueError:
-                logging.debug(f"No convolution #{i}")
-
-    def reset(self):
-        """Reset read offset to beginning of weight array."""
-        self.offset = 0
-
-class BoundBox:
-    """
-    Represents a bounding box detection with class probabilities.
-
-    Stores bounding box coordinates along with objectness score and
-    class probabilities for all COCO classes. Provides cached access
-    to predicted label and confidence score.
-
-    Attributes:
-        xmin (float): Left coordinate (normalized 0-1 or pixel coordinates)
-        ymin (float): Top coordinate
-        xmax (float): Right coordinate
-        ymax (float): Bottom coordinate
-        objness (float): Objectness score (probability of containing an object)
-        classes (np.ndarray): Array of class probabilities (80 classes for COCO)
-        label (int): Cached predicted class index (-1 until computed)
-        score (float): Cached confidence score (-1 until computed)
-
-    Example:
-        >>> box = BoundBox(0.1, 0.2, 0.5, 0.8, 0.95, class_probs)
-        >>> print(f"Class: {box.get_label()}, Confidence: {box.get_score():.2f}")
-        >>> print(f"Box: ({box.xmin}, {box.ymin}, {box.xmax}, {box.ymax})")
-    """
-    def __init__(self, xmin, ymin, xmax, ymax, objness = None, classes = None):
-        """
-        Initialize bounding box with coordinates and predictions.
-
-        Args:
-            xmin (float): Left edge of box
-            ymin (float): Top edge of box
-            xmax (float): Right edge of box
-            ymax (float): Bottom edge of box
-            objness (float, optional): Objectness score
-            classes (np.ndarray, optional): Class probability array
-        """
-        self.xmin = xmin
-        self.ymin = ymin
-        self.xmax = xmax
-        self.ymax = ymax
-
-        self.objness = objness
-        self.classes = classes
-
-        self.label = -1
-        self.score = -1
-
-    def get_label(self):
-        """
-        Get predicted class label (index with highest probability).
-
-        Returns:
-            int: Class index (0-79 for COCO dataset)
-
-        Note:
-            Result is cached for efficiency.
-        """
-        if self.label == -1:
-            self.label = np.argmax(self.classes)
-
-        return self.label
-
-    def get_score(self):
-        """
-        Get confidence score for predicted class.
-
-        Returns:
-            float: Confidence score (0-1)
-
-        Note:
-            Combines objectness and class probability.
-            Result is cached for efficiency.
-        """
-        if self.score == -1:
-            self.score = self.classes[self.get_label()]
-
-        return self.score
-
-def _conv_block(inp, convs, skip=True):
-    """
-    Create a convolutional block for YOLOv3 architecture.
-
-    This function constructs a sequence of convolutional layers with optional
-    batch normalization and LeakyReLU activation. It can also add a skip
-    connection (residual connection) for better gradient flow.
-
-    Args:
-        inp: Input tensor from previous layer
-        convs (list): List of dictionaries defining convolution layers.
-            Each dict should contain:
-                - filter (int): Number of filters
-                - kernel (int): Kernel size
-                - stride (int): Stride value
-                - bnorm (bool): Whether to use batch normalization
-                - leaky (bool): Whether to use LeakyReLU activation
-                - layer_idx (int): Layer index for naming
-        skip (bool): If True, add skip connection. Default: True
-
-    Returns:
-        Tensor: Output tensor after applying all convolutions
-
-    Note:
-        Uses peculiar padding (left and top) to match Darknet implementation.
-    """
-    x = inp
-    count = 0
-
-    for conv in convs:
-        if count == (len(convs) - 2) and skip:
-            skip_connection = x
-        count += 1
-
-        if conv['stride'] > 1: x = ZeroPadding2D(((1,0),(1,0)))(x) # peculiar padding as darknet prefer left and top
-        x = Conv2D(conv['filter'],
-                   conv['kernel'],
-                   strides=conv['stride'],
-                   padding='valid' if conv['stride'] > 1 else 'same', # peculiar padding as darknet prefer left and top
-                   name='conv_' + str(conv['layer_idx']),
-                   use_bias=False if conv['bnorm'] else True)(x)
-        if conv['bnorm']: x = BatchNormalization(epsilon=0.001, name='bnorm_' + str(conv['layer_idx']))(x)
-        if conv['leaky']: x = LeakyReLU(alpha=0.1, name='leaky_' + str(conv['layer_idx']))(x)
-
-    return Add()([skip_connection, x]) if skip else x
-
-def _interval_overlap(interval_a, interval_b):
-    """
-    Calculate the overlap between two 1D intervals.
-
-    Used as a helper function for computing Intersection over Union (IoU)
-    between bounding boxes.
-
-    Args:
-        interval_a (tuple): First interval as (start, end)
-        interval_b (tuple): Second interval as (start, end)
-
-    Returns:
-        float: Length of overlap between intervals. Returns 0 if no overlap.
-
-    Example:
-        >>> _interval_overlap((0, 10), (5, 15))
-        5
-        >>> _interval_overlap((0, 5), (10, 15))
-        0
-    """
-    x1, x2 = interval_a
-    x3, x4 = interval_b
-
-    if x3 < x1:
-        if x4 < x1:
-            return 0
-        else:
-            return min(x2,x4) - x1
-    else:
-        if x2 < x3:
-             return 0
-        else:
-            return min(x2,x4) - x3          
-
-def _sigmoid(x):
-    """
-    Apply sigmoid activation function.
-
-    Args:
-        x (np.ndarray or float): Input value(s)
-
-    Returns:
-        np.ndarray or float: Sigmoid activation output in range (0, 1)
-
-    Example:
-        >>> _sigmoid(0)
-        0.5
-        >>> _sigmoid(np.array([0, 1, -1]))
-        array([0.5, 0.73105858, 0.26894142])
-    """
-    return 1. / (1. + np.exp(-x))
-
-def bbox_iou(box1, box2):
-    """
-    Calculate Intersection over Union (IoU) between two bounding boxes.
-
-    IoU is a measure of overlap between two bounding boxes, commonly used
-    in object detection for Non-Maximum Suppression (NMS).
-
-    Args:
-        box1 (BoundBox): First bounding box
-        box2 (BoundBox): Second bounding box
-
-    Returns:
-        float: IoU score between 0 and 1.
-            - 0 means no overlap
-            - 1 means perfect overlap
-
-    Formula:
-        IoU = Area of Intersection / Area of Union
-
-    Example:
-        >>> box1 = BoundBox(0, 0, 10, 10)
-        >>> box2 = BoundBox(5, 5, 15, 15)
-        >>> iou = bbox_iou(box1, box2)
-        >>> print(f"IoU: {iou:.2f}")
-    """
-    intersect_w = _interval_overlap([box1.xmin, box1.xmax], [box2.xmin, box2.xmax])
-    intersect_h = _interval_overlap([box1.ymin, box1.ymax], [box2.ymin, box2.ymax])
-
-    intersect = intersect_w * intersect_h
-
-    w1, h1 = box1.xmax-box1.xmin, box1.ymax-box1.ymin
-    w2, h2 = box2.xmax-box2.xmin, box2.ymax-box2.ymin
-
-    union = w1*h1 + w2*h2 - intersect
-
-    return float(intersect) / union
-
-def make_yolov3_model():
-    """
-    Construct the YOLOv3 (Darknet-53) neural network architecture.
-
-    This function builds the complete YOLOv3 model with 106 convolutional layers
-    organized in a Darknet-53 backbone with three detection heads for different
-    scales (small, medium, and large objects).
-
-    Returns:
-        keras.Model: YOLOv3 model with three outputs:
-            - yolo_82: Large object detection (13x13 grid)
-            - yolo_94: Medium object detection (26x26 grid)
-            - yolo_106: Small object detection (52x52 grid)
-
-    Architecture:
-        - Input: Variable size RGB image (H, W, 3)
-        - Backbone: Darknet-53 (53 convolutional layers)
-        - Detection heads: 3 scales with different receptive fields
-        - Output: 3 tensors for multi-scale detection
-
-    Note:
-        Model requires pre-trained weights from COCO dataset.
-        Use WeightReader.load_weights() to load weights after creation.
-
-    Example:
-        >>> model = make_yolov3_model()
-        >>> print(f"Total parameters: {model.count_params():,}")
-        >>> weight_reader = WeightReader('yolov3.weights')
-        >>> weight_reader.load_weights(model)
-    """
-    input_image = Input(shape=(None, None, 3))
-
-    # Layer  0 => 4
-    x = _conv_block(input_image, [{'filter': 32, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 0},
-                                  {'filter': 64, 'kernel': 3, 'stride': 2, 'bnorm': True, 'leaky': True, 'layer_idx': 1},
-                                  {'filter': 32, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 2},
-                                  {'filter': 64, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 3}])
-
-    # Layer  5 => 8
-    x = _conv_block(x, [{'filter': 128, 'kernel': 3, 'stride': 2, 'bnorm': True, 'leaky': True, 'layer_idx': 5},
-                        {'filter':  64, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 6},
-                        {'filter': 128, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 7}])
-
-    # Layer  9 => 11
-    x = _conv_block(x, [{'filter':  64, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 9},
-                        {'filter': 128, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 10}])
-
-    # Layer 12 => 15
-    x = _conv_block(x, [{'filter': 256, 'kernel': 3, 'stride': 2, 'bnorm': True, 'leaky': True, 'layer_idx': 12},
-                        {'filter': 128, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 13},
-                        {'filter': 256, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 14}])
-
-    # Layer 16 => 36
-    for i in range(7):
-        x = _conv_block(x, [{'filter': 128, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 16+i*3},
-                            {'filter': 256, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 17+i*3}])
-        
-    skip_36 = x
-        
-    # Layer 37 => 40
-    x = _conv_block(x, [{'filter': 512, 'kernel': 3, 'stride': 2, 'bnorm': True, 'leaky': True, 'layer_idx': 37},
-                        {'filter': 256, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 38},
-                        {'filter': 512, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 39}])
-
-    # Layer 41 => 61
-    for i in range(7):
-        x = _conv_block(x, [{'filter': 256, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 41+i*3},
-                            {'filter': 512, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 42+i*3}])
-        
-    skip_61 = x
-        
-    # Layer 62 => 65
-    x = _conv_block(x, [{'filter': 1024, 'kernel': 3, 'stride': 2, 'bnorm': True, 'leaky': True, 'layer_idx': 62},
-                        {'filter':  512, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 63},
-                        {'filter': 1024, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 64}])
-
-    # Layer 66 => 74
-    for i in range(3):
-        x = _conv_block(x, [{'filter':  512, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 66+i*3},
-                            {'filter': 1024, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 67+i*3}])
-        
-    # Layer 75 => 79
-    x = _conv_block(x, [{'filter':  512, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 75},
-                        {'filter': 1024, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 76},
-                        {'filter':  512, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 77},
-                        {'filter': 1024, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 78},
-                        {'filter':  512, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 79}], skip=False)
-
-    # Layer 80 => 82
-    yolo_82 = _conv_block(x, [{'filter': 1024, 'kernel': 3, 'stride': 1, 'bnorm': True,  'leaky': True,  'layer_idx': 80},
-                              {'filter':  255, 'kernel': 1, 'stride': 1, 'bnorm': False, 'leaky': False, 'layer_idx': 81}], skip=False)
-
-    # Layer 83 => 86
-    x = _conv_block(x, [{'filter': 256, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 84}], skip=False)
-    x = UpSampling2D(2)(x)
-    x = Concatenate()([x, skip_61])
-
-    # Layer 87 => 91
-    x = _conv_block(x, [{'filter': 256, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 87},
-                        {'filter': 512, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 88},
-                        {'filter': 256, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 89},
-                        {'filter': 512, 'kernel': 3, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 90},
-                        {'filter': 256, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True, 'layer_idx': 91}], skip=False)
-
-    # Layer 92 => 94
-    yolo_94 = _conv_block(x, [{'filter': 512, 'kernel': 3, 'stride': 1, 'bnorm': True,  'leaky': True,  'layer_idx': 92},
-                              {'filter': 255, 'kernel': 1, 'stride': 1, 'bnorm': False, 'leaky': False, 'layer_idx': 93}], skip=False)
-
-    # Layer 95 => 98
-    x = _conv_block(x, [{'filter': 128, 'kernel': 1, 'stride': 1, 'bnorm': True, 'leaky': True,   'layer_idx': 96}], skip=False)
-    x = UpSampling2D(2)(x)
-    x = Concatenate()([x, skip_36])
-
-    # Layer 99 => 106
-    yolo_106 = _conv_block(x, [{'filter': 128, 'kernel': 1, 'stride': 1, 'bnorm': True,  'leaky': True,  'layer_idx': 99},
-                               {'filter': 256, 'kernel': 3, 'stride': 1, 'bnorm': True,  'leaky': True,  'layer_idx': 100},
-                               {'filter': 128, 'kernel': 1, 'stride': 1, 'bnorm': True,  'leaky': True,  'layer_idx': 101},
-                               {'filter': 256, 'kernel': 3, 'stride': 1, 'bnorm': True,  'leaky': True,  'layer_idx': 102},
-                               {'filter': 128, 'kernel': 1, 'stride': 1, 'bnorm': True,  'leaky': True,  'layer_idx': 103},
-                               {'filter': 256, 'kernel': 3, 'stride': 1, 'bnorm': True,  'leaky': True,  'layer_idx': 104},
-                               {'filter': 255, 'kernel': 1, 'stride': 1, 'bnorm': False, 'leaky': False, 'layer_idx': 105}], skip=False)
-
-    model = Model(input_image, [yolo_82, yolo_94, yolo_106])    
-    return model
-
-def preprocess_input(image, net_h, net_w):
-    """
-    Preprocess input image for YOLOv3 detection.
-
-    Resizes image to network input size while maintaining aspect ratio using
-    letterboxing (padding with gray). Converts BGR to RGB and normalizes
-    pixel values to [0, 1].
-
-    Args:
-        image (np.ndarray): Input image in BGR format (H, W, 3)
-        net_h (int): Network input height (typically 416)
-        net_w (int): Network input width (typically 416)
-
-    Returns:
-        np.ndarray: Preprocessed image with shape (1, net_h, net_w, 3)
-            - Normalized to [0, 1]
-            - RGB format
-            - Letterboxed to maintain aspect ratio
-            - Batch dimension added
-
-    Example:
-        >>> frame = cv2.imread('image.jpg')  # (720, 1280, 3)
-        >>> preprocessed = preprocess_input(frame, 416, 416)
-        >>> print(preprocessed.shape)
-        (1, 416, 416, 3)
-    """
-    new_h, new_w, _ = image.shape
-
-    # determine the new size of the image
-    if (float(net_w)/new_w) < (float(net_h)/new_h):
-        new_h = (new_h * net_w)/new_w
-        new_w = net_w
-    else:
-        new_w = (new_w * net_h)/new_h
-        new_h = net_h
-
-    # resize the image to the new size
-    resized = cv2.resize(image[:,:,::-1]/255., (int(new_w), int(new_h)))
-
-    # embed the image into the standard letter box
-    new_image = np.ones((net_h, net_w, 3)) * 0.5
-    new_image[int((net_h-new_h)//2):int((net_h+new_h)//2), int((net_w-new_w)//2):int((net_w+new_w)//2), :] = resized
-    new_image = np.expand_dims(new_image, 0)
-
-    return new_image
-
-def decode_netout(netout, anchors, obj_thresh, nms_thresh, net_h, net_w):
-    """
-    Decode YOLOv3 network output into bounding boxes.
-
-    Converts raw network predictions into interpretable bounding boxes with
-    class probabilities. Applies objectness threshold to filter low-confidence
-    detections.
-
-    Args:
-        netout (np.ndarray): Network output tensor (grid_h, grid_w, 255)
-        anchors (list): Anchor box dimensions for this scale [w1,h1, w2,h2, w3,h3]
-        obj_thresh (float): Objectness threshold (0-1) for filtering detections
-        nms_thresh (float): NMS threshold (not used in this function)
-        net_h (int): Network input height
-        net_w (int): Network input width
-
-    Returns:
-        list[BoundBox]: List of detected bounding boxes with class probabilities
-
-    Note:
-        Each grid cell predicts 3 bounding boxes (one per anchor).
-        Output shape 255 = 3 * (5 + 80) where:
-            - 5 = (x, y, w, h, objectness)
-            - 80 = class probabilities for COCO dataset
-
-    Example:
-        >>> yolos = model.predict(preprocessed_image)
-        >>> boxes = decode_netout(yolos[0][0], anchors[0], 0.5, 0.45, 416, 416)
-        >>> print(f"Detected {len(boxes)} boxes")
-    """
-    grid_h, grid_w = netout.shape[:2]
-    nb_box = 3
-    netout = netout.reshape((grid_h, grid_w, nb_box, -1))
-    nb_class = netout.shape[-1] - 5
-
-    boxes = []
-
-    netout[..., :2]  = _sigmoid(netout[..., :2])
-    netout[..., 4:]  = _sigmoid(netout[..., 4:])
-    netout[..., 5:]  = netout[..., 4][..., np.newaxis] * netout[..., 5:]
-    netout[..., 5:] *= netout[..., 5:] > obj_thresh
-
-    for i in range(grid_h*grid_w):
-        row = i // grid_w  # Integer division
-        col = i % grid_w
-        
-        for b in range(nb_box):
-            # 4th element is objectness score
-            objectness = netout[int(row)][int(col)][b][4]
-            #objectness = netout[..., :4]
-            
-            if(objectness.all() <= obj_thresh): continue
-            
-            # first 4 elements are x, y, w, and h
-            x, y, w, h = netout[int(row)][int(col)][b][:4]
-
-            x = (col + x) / grid_w # center position, unit: image width
-            y = (row + y) / grid_h # center position, unit: image height
-            w = anchors[2 * b + 0] * np.exp(w) / net_w # unit: image width
-            h = anchors[2 * b + 1] * np.exp(h) / net_h # unit: image height  
-            
-            # last elements are class probabilities
-            classes = netout[int(row)][col][b][5:]
-            
-            box = BoundBox(x-w/2, y-h/2, x+w/2, y+h/2, objectness, classes)
-            #box = BoundBox(x-w/2, y-h/2, x+w/2, y+h/2, None, classes)
-
-            boxes.append(box)
-
-    return boxes
-
-def correct_yolo_boxes(boxes, image_h, image_w, net_h, net_w):
-    """
-    Correct bounding box coordinates to match original image dimensions.
-
-    YOLOv3 predictions are relative to the letterboxed image. This function
-    transforms coordinates back to the original image space, accounting for
-    the letterboxing (padding) applied during preprocessing.
-
-    Args:
-        boxes (list[BoundBox]): List of bounding boxes to correct (modified in-place)
-        image_h (int): Original image height
-        image_w (int): Original image width
-        net_h (int): Network input height (e.g., 416)
-        net_w (int): Network input width (e.g., 416)
-
-    Returns:
-        None: Modifies boxes in-place
-
-    Note:
-        This function accounts for the aspect-ratio-preserving resize and
-        letterboxing applied in preprocess_input().
-
-    Example:
-        >>> boxes = decode_netout(...)  # boxes in network coordinates
-        >>> correct_yolo_boxes(boxes, 720, 1280, 416, 416)
-        >>> # boxes now in original image coordinates (0-720, 0-1280)
-    """
-    if (float(net_w)/image_w) < (float(net_h)/image_h):
-        new_w = net_w
-        new_h = (image_h*net_w)/image_w
-    else:
-        new_h = net_w
-        new_w = (image_w*net_h)/image_h
-
-    for i in range(len(boxes)):
-        x_offset, x_scale = (net_w - new_w)/2./net_w, float(new_w)/net_w
-        y_offset, y_scale = (net_h - new_h)/2./net_h, float(new_h)/net_h
-
-        boxes[i].xmin = int((boxes[i].xmin - x_offset) / x_scale * image_w)
-        boxes[i].xmax = int((boxes[i].xmax - x_offset) / x_scale * image_w)
-        boxes[i].ymin = int((boxes[i].ymin - y_offset) / y_scale * image_h)
-        boxes[i].ymax = int((boxes[i].ymax - y_offset) / y_scale * image_h)
-        
-def do_nms(boxes, nms_thresh):
-    """
-    Apply Non-Maximum Suppression (NMS) to remove duplicate detections.
-
-    NMS suppresses overlapping bounding boxes, keeping only the highest
-    confidence detection for each object. Applied per-class to handle
-    multi-class detection scenarios.
-
-    Args:
-        boxes (list[BoundBox]): List of bounding boxes (modified in-place)
-        nms_thresh (float): IoU threshold (0-1) for suppression.
-            - Higher values = more aggressive suppression
-            - Typical value: 0.45
-
-    Returns:
-        None: Modifies boxes in-place by zeroing out class probabilities
-              of suppressed boxes
-
-    Algorithm:
-        For each class:
-        1. Sort boxes by class probability (descending)
-        2. For each box, suppress all lower-scoring boxes with IoU > threshold
-        3. Suppressed boxes have their class probability set to 0
-
-    Example:
-        >>> boxes = decode_netout(...)
-        >>> print(f"Before NMS: {len(boxes)} boxes")
-        >>> do_nms(boxes, 0.45)
-        >>> # Filter out suppressed boxes
-        >>> boxes = [box for box in boxes if box.get_score() > 0]
-        >>> print(f"After NMS: {len(boxes)} boxes")
-    """
-    if len(boxes) > 0:
-        nb_class = len(boxes[0].classes)
-    else:
-        return
-
-    for c in range(nb_class):
-        sorted_indices = np.argsort([-box.classes[c] for box in boxes])
-
-        for i in range(len(sorted_indices)):
-            index_i = sorted_indices[i]
-
-            if boxes[index_i].classes[c] == 0: continue
-
-            for j in range(i+1, len(sorted_indices)):
-                index_j = sorted_indices[j]
-
-                if bbox_iou(boxes[index_i], boxes[index_j]) >= nms_thresh:
-                    boxes[index_j].classes[c] = 0
-                    
-def draw_boxes(image, boxes, line, labels, obj_thresh, dcnt):
-    """
-    Draw bounding boxes on image and detect traffic violations.
-
-    Draws detection boxes on the image with different colors:
-    - Green: Compliant vehicles (not crossing line)
-    - Red: Violation detected (crossing traffic signal line)
-
-    Also draws the traffic signal line and saves violation snapshots.
-
-    Args:
-        image (np.ndarray): Input image (will be modified in-place)
-        boxes (list[BoundBox]): List of detected bounding boxes
-        line (list): Traffic line coordinates [(x1, y1), (x2, y2)]
-        labels (list[str]): Class label names (COCO labels)
-        obj_thresh (float): Confidence threshold for displaying boxes
-        dcnt (int): Frame counter for naming violation snapshots
-
-    Returns:
-        np.ndarray: Annotated image with boxes and line drawn
-
-    Side Effects:
-        - Saves violation snapshots to config.DETECTED_IMAGES_DIR
-        - Displays violation window with cv2.imshow (if not headless)
-
-    Violation Detection:
-        A violation is detected when any edge of a vehicle's bounding box
-        intersects with the traffic signal line, indicating the vehicle
-        has crossed the red signal.
-
-    Example:
-        >>> frame = cv2.imread('traffic.jpg')
-        >>> boxes = [...]  # After detection and NMS
-        >>> line = [(100, 200), (500, 200)]  # Horizontal line
-        >>> annotated = draw_boxes(frame, boxes, line, LABELS, 0.5, 1)
-        >>> cv2.imwrite('annotated.jpg', annotated)
-    """
-    logging.debug(f"Drawing boxes with line: {line}")
-
-    for box in boxes:
-        label_str = ''
-        label = -1
-
-        for i in range(len(labels)):
-            if box.classes[i] > obj_thresh:
-                label_str += labels[i]
-                label = i
-                logging.debug(f"{labels[i]}: {box.classes[i]*100:.2f}%")
-                logging.debug(f"Line: ({line[0][0]}, {line[0][1]}) ({line[1][0]}, {line[1][1]})")
-                logging.debug(f"Box: ({box.xmin}, {box.ymin}) ({box.xmax}, {box.ymax})")
-
-        if label >= 0:
-            # Check if bounding box intersects with traffic signal line
-            bbox_min = (box.xmin, box.ymin)
-            bbox_max = (box.xmax, box.ymax)
-            is_violation = bounding_box_intersects_line(bbox_min, bbox_max, line[0], line[1])
-
-            logging.debug(f"Violation detected: {is_violation}")
-
-            cv2.line(image, line[0], line[1], (255, 0, 0), 3)
-
-            if is_violation:
-                # Red box for violation
-                cv2.rectangle(image, (box.xmin,box.ymin), (box.xmax,box.ymax), (255,0,0), 3)
-                cimg = image[box.ymin:box.ymax, box.xmin:box.xmax]
-                cv2.imshow("violation", cimg)
-                cv2.waitKey(5)
-                violation_path = config.DETECTED_IMAGES_DIR / f"violation_{dcnt}.jpg"
-                cv2.imwrite(str(violation_path), cimg)
-                dcnt = dcnt+1
-            else:
-                # Green box for compliant vehicle
-                cv2.rectangle(image, (box.xmin,box.ymin), (box.xmax,box.ymax), (0,255,0), 3)
-
-            cv2.putText(image, 
-                        label_str + ' ' + str(round(box.get_score(), 2)), 
-                        (box.xmin, box.ymin - 13), 
-                        cv2.FONT_HERSHEY_SIMPLEX, 
-                        1e-3 * image.shape[0], 
-                        (0,255,0), 2)
-        
-    return image
-
-# ============================================================================
-# YOLOv3 Violation Detector with Lazy Loading
-# ============================================================================
-
-class ViolationDetector:
-    """
-    Lazy-loading wrapper for YOLOv3 traffic violation detection model.
-
-    The model is not loaded until the first prediction is made,
-    improving startup time for non-processing operations.
-
-    Attributes:
-        _model: Cached Keras model (None until first use)
-        _load_time: Time taken to load model (seconds)
-
-    Performance:
-        - First prediction: ~30s (model loading + inference)
-        - Subsequent predictions: ~0.1s (inference only)
-
-    Example:
-        >>> detector = ViolationDetector()
-        >>> # Model not loaded yet
-        >>> image = preprocess_input(frame, config.NET_H, config.NET_W)
-        >>> predictions = detector.predict(image)  # Model loads here
-        >>> predictions2 = detector.predict(image2)  # Uses cached model
-    """
-
-    def __init__(self):
-        """
-        Initialize detector without loading model.
-
-        The model will be loaded lazily on first predict() call.
-        """
-        self._model = None
-        self._load_time = None
-        logging.info("ViolationDetector initialized (lazy loading enabled)")
-
-    @property
-    def model(self):
-        """
-        Get the YOLOv3 model, loading it on first access.
-
-        This property implements lazy loading - the expensive model
-        loading operation only happens on first access.
-
-        Returns:
-            keras.Model: Loaded YOLOv3 model ready for inference
-
-        Raises:
-            FileNotFoundError: If weights file doesn't exist
-            RuntimeError: If model loading fails
-        """
-        if self._model is None:
-            self._load_model()
-        return self._model
-
-    def _load_model(self):
-        """
-        Internal method to load YOLOv3 model and weights.
-
-        Raises:
-            FileNotFoundError: If WEIGHTS_PATH doesn't exist
-            RuntimeError: If model creation or weight loading fails
-        """
-        logging.info("Loading YOLOv3 model...")
-        start_time = time.time()
-
-        try:
-            # Verify weights file exists
-            if not config.WEIGHTS_PATH.exists():
-                raise FileNotFoundError(
-                    f"YOLOv3 weights not found at {config.WEIGHTS_PATH}. "
-                    "Please download from: https://pjreddie.com/media/files/yolov3.weights"
-                )
-
-            # Create model architecture
-            logging.info("Building YOLOv3 architecture...")
-            self._model = make_yolov3_model()
-
-            # Load pre-trained weights
-            logging.info(f"Loading weights from {config.WEIGHTS_PATH}...")
-            weight_reader = WeightReader(str(config.WEIGHTS_PATH))
-            weight_reader.load_weights(self._model)
-
-            # Calculate load time
-            self._load_time = time.time() - start_time
-
-            logging.info(f"✅ Model loaded successfully in {self._load_time:.2f}s")
-            logging.info(f"Model parameters: {self._model.count_params():,}")
-
-        except Exception as e:
-            logging.error(f"Failed to load YOLOv3 model: {e}")
-            raise RuntimeError(f"Model loading failed: {e}") from e
-
-    def predict(self, image):
-        """
-        Run object detection on preprocessed image.
-
-        Args:
-            image (np.ndarray): Preprocessed image with shape (1, NET_H, NET_W, 3)
-
-        Returns:
-            list: YOLOv3 predictions for three scales
-                  [output_1, output_2, output_3]
-
-        Raises:
-            ValueError: If image has wrong shape
-            RuntimeError: If prediction fails
-
-        Example:
-            >>> preprocessed = preprocess_input(frame, config.NET_H, config.NET_W)
-            >>> predictions = detector.predict(preprocessed)
-            >>> boxes = decode_netout(predictions[0], ...)
-        """
-        # Validate input shape
-        expected_shape = (1, config.NET_H, config.NET_W, 3)
-        if image.shape != expected_shape:
-            raise ValueError(
-                f"Expected image shape {expected_shape}, got {image.shape}"
-            )
-
-        try:
-            # This will trigger lazy loading if needed
-            return self.model.predict(image, verbose=0)
-        except Exception as e:
-            logging.error(f"Prediction failed: {e}")
-            raise RuntimeError(f"Detection failed: {e}") from e
-
-    @property
-    def is_loaded(self):
-        """Check if model is currently loaded in memory."""
-        return self._model is not None
-
-    def unload(self):
-        """
-        Unload model from memory to free resources.
-
-        Useful for long-running applications that process videos
-        in batches with idle time between batches.
-        """
-        if self._model is not None:
-            logging.info("Unloading model from memory")
-            del self._model
-            self._model = None
-
-            # Force garbage collection
-            import gc
-            gc.collect()
-
-
-# ============================================================================
-# Global Detector Instance (Lazy Loading)
-# ============================================================================
-
-# Create global detector instance (model will load on first predict() call)
-detector = ViolationDetector()
-
-# ============================================================================
-# Legacy Global Variables (for backward compatibility)
-# ============================================================================
-
-# OLD CODE - DEPRECATED (kept for backward compatibility)
-# These will be removed in future versions
-# Use config.py values instead
-
-# weights_path = "G:/Traffic Violation Detection/yolov3.weights"  # DEPRECATED
-# net_h, net_w = 416, 416  # DEPRECATED - use config.NET_H, config.NET_W
-# obj_thresh, nms_thresh = 0.5, 0.45  # DEPRECATED - use config.OBJ_THRESH, config.NMS_THRESH
-# anchors = [[116,90,  156,198,  373,326],  [30,61, 62,45,  59,119], [10,13,  16,30,  33,23]]  # DEPRECATED - use config.ANCHORS
-# labels = [...]  # DEPRECATED - use config.LABELS
-
-# OLD MODEL LOADING - COMMENTED OUT (now using lazy loading via detector)
-# yolov3 = make_yolov3_model()
-# weight_reader = WeightReader(weights_path)
-# weight_reader.load_weights(yolov3)
+"""
+DEPRECATED: Legacy object_detection module
+
+This module is deprecated and maintained only for backward compatibility.
+Please update your imports to use the new modular structure:
+
+OLD:
+    from object_detection import detector, preprocess_input, draw_boxes
+
+NEW:
+    from detection import detector, preprocess_input, draw_boxes
+    from models import make_yolov3_model, WeightReader
+
+This file will be removed in a future version.
+"""
+
+import warnings
+
+# Show deprecation warning
+warnings.warn(
+    "object_detection module is deprecated. "
+    "Use 'from detection import ...' and 'from models import ...' instead. "
+    "See module docstring for migration guide.",
+    DeprecationWarning,
+    stacklevel=2
+)
+
+# Re-export everything from new modular structure for backward compatibility
+from models import make_yolov3_model, WeightReader
+from detection import (
+    ViolationDetector,
+    detector,
+    preprocess_input,
+    decode_netout,
+    correct_yolo_boxes,
+    do_nms,
+    BoundBox,
+    bbox_iou,
+    draw_boxes
+)
+
+__all__ = [
+    # Model components
+    'make_yolov3_model',
+    'WeightReader',
+    # Detection components
+    'ViolationDetector',
+    'detector',
+    'preprocess_input',
+    'decode_netout',
+    'correct_yolo_boxes',
+    'do_nms',
+    # Violation checking
+    'BoundBox',
+    'bbox_iou',
+    'draw_boxes',
+]
